Restructure of packages to make them per feature: Case 1 OpenCV / Case 2 Camera2 closes #1 Presenter introduced in Case 1 OpenCV closes #2 FileHelper to decouple file manipulation
