+++ app/src/main/AndroidManifest.xml
@@ -48,7 +48,7 @@
         android:theme="@android:style/Theme.NoTitleBar.Fullscreen" >
 
         <activity
-            android:name="com.raulh82vlc.image_recognition_sample.ui.activity.FDOpenCVActivity"
+            android:name="com.raulh82vlc.image_recognition_sample.opencv.ui.FDOpenCVActivity"
             android:label="@string/OpenCV"
             android:hardwareAccelerated="true"
             android:screenOrientation="landscape"
@@ -57,7 +57,7 @@
         </activity>
 
         <activity
-            android:name="com.raulh82vlc.image_recognition_sample.ui.activity.FDCamera2Activity"
+            android:name="com.raulh82vlc.image_recognition_sample.camera2.ui.FDCamera2Activity"
             android:label="@string/Camera2"
             android:hardwareAccelerated="true"
             android:screenOrientation="portrait"
@@ -66,7 +66,7 @@
         </activity>
 
         <activity
-            android:name="com.raulh82vlc.image_recognition_sample.ui.activity.MainActivity"
+            android:name="com.raulh82vlc.image_recognition_sample.ui.MainActivity"
             android:label="@string/app_name"
             android:launchMode="singleTask"
             android:configChanges="keyboardHidden|orientation">
