+++ app/src/main/res/layout/activity_face_detection.xml
@@ -25,7 +25,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"/>
 
-    <com.raulh82vlc.image_recognition_sample.ui.widgets.FaceDrawer
+    <com.raulh82vlc.image_recognition_sample.camera2.render.FaceDrawer
         android:id="@+id/face_drawer"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
